@@ -46,36 +46,42 @@
     "test": "xo"
   },
   "dependencies": {
-    "@types/node": "^20.9.0",
-    "chalk": "^2.4.1",
-    "clipboardy": "^1.2.3",
-    "commander": "^11.0.0",
+    "chalk": "4.1.0",
+    "clipboardy": "2.3.0",
+    "commander": "^11.1.0",
     "enquirer": "^2.4.1",
     "env-paths": "^3.0.0",
-    "marked": "^9.0.3",
-    "marked-terminal": "^6.0.0",
-    "meow": "^5.0.0",
-    "nanoid": "^3.3.6",
+    "marked": "^10.0.0",
+    "marked-terminal": "^6.1.0",
+    "meow": "^12.1.1",
+    "nanoid": "3",
     "signale": "1.4.0",
     "tmp": "^0.2.1",
     "typescript": "^5.2.2",
-    "update-notifier": "^2.5.0"
+    "update-notifier": "5.0.1"
   },
   "devDependencies": {
+    "@types/chalk": "^2.2.0",
+    "@types/marked-terminal": "^6.0.1",
+    "@types/node": "^20.9.1",
+    "@types/signale": "^1.4.7",
+    "@types/tmp": "^0.2.6",
+    "@types/update-notifier": "^6.0.8",
     "@typescript-eslint/eslint-plugin": "^6.11.0",
     "@typescript-eslint/parser": "^6.11.0",
-    "eslint": "^8.53.0",
+    "eslint": "^8.54.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-formatter-pretty": "^5.0.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^5.0.0",
-    "prettier": "^3.0.3",
+    "eslint-formatter-pretty": "^6.0.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "prettier": "^3.1.0",
     "semver": "^7.5.4",
     "xo": "*"
   },
   "xo": {
     "space": 2
-  }
+  },
+  "packageManager": "yarn@4.0.2"
 }
