@@ -6,7 +6,6 @@ import chalk from 'chalk'
 import pkg from '../package.json'
 
 // for self-documentation
-type ChalkColorString = string
 type PriorityLevel = 1 | 2 | 3
 
 interface IConfig {
@@ -14,14 +13,18 @@ interface IConfig {
   displayCompleteTasks: boolean
   displayProgressOverview: boolean
   displayWarnings: boolean
-  priorities: Record<PriorityLevel, ChalkColorString>
-  highlightTitle: string
+  priorities: Record<PriorityLevel, Function>
+  highlightTitle: Function
   enableCopyID: boolean
   eventBoard: string
+  defaultBoard: string
+  editor: string
+  suspiciousDuration: number
 }
 
 const { default: defaultConfig } = pkg.configuration
 // TODO: https://github.com/sindresorhus/env-paths
+// TODO: Have a fallback on ~/.taskbook
 const CONFIG_DIR = '.config'
 
 class Config {
@@ -29,8 +32,12 @@ class Config {
   private _configFile: string
 
   constructor() {
-    // supprot for cross platform standard paths
-    this._configPath = path.join(os.homedir(), CONFIG_DIR, pkg.name)
+    // old taskbook was using this and it has far more users than my package.
+    // Pretty easy to be compatible, no need to force users into migration
+    const legacyConfigPath = path.join(os.homedir(), '.taskbook')
+    if (fs.existsSync(legacyConfigPath)) this._configPath = path.join(legacyConfigPath)
+    else this._configPath = path.join(os.homedir(), CONFIG_DIR, pkg.name)
+
     this._configFile = path.join(this._configPath, 'config.json')
 
     this._ensureConfigFile()
@@ -50,15 +57,19 @@ class Config {
     const config = JSON.parse(content)
 
     return {
-      // ~/.taskbook/config.json
-      ...config,
-
+      // defaults
       taskbookDirectory: this._configPath,
       // TODO: support in package.json and/or environment
       // constants
       // NOTE: will have to do with theming if implemented
-      priorities: { 2: 'yellow', 3: 'red' },
+      priorities: { 2: chalk.underline.yellow, 3: chalk.underline.red },
       highlightTitle: chalk.bold.cyan,
+      defaultBoard: 'My Board',
+      editor: process.env.EDITOR || 'vi',
+      suspiciousDuration: 3 /* hours */,
+
+      // ~/.config/taskbook/config.json
+      ...config,
 
       /**
        * TODO: not supported yet (would really benefit from a flat config)
@@ -67,7 +78,6 @@ class Config {
        * prioritiesLabels: ['p:1', 'p:2', 'p:3'],
        * boardPrefix: '@',
        * tagPrefix: '+',
-       * suspiciousDuration: 3 // hours
        */
     }
   }
